requirements.txt
pandas==1.4.2
mysql-connector-python==8.0.32
google-api-python-client==2.58.0
requests==2.28.1
isodate==0.6.1